+++ kedro_mlflow/__init__.py
@@ -1,6 +1,6 @@
-"""kedro-mlflow plugin constants
-"""
-__version__ = "0.12.2"
+"""kedro-mlflow plugin constants"""
+
+__version__ = "0.13.1"
 
 import logging